Unlock mobile network settings when airplane mode is on
In many cases, it can be helpful to open mobile network settings when
airplane mode is enabled. A few cases come to mind:
  - Checking data usage to see if it's safe to re-enable mobile data
  - Disabling mobile data to avoid overage charges before reconnecting
  - Changing settings while exclusively using Wi-Fi calling
  - Updating data limit before reconnecting to avoid overage charges
  - Fixing or updating APNs before reconnecting

Most of the options exposed in mobile network settings do not require an
active connection to work, so there's no reason to prevent the user from
accessing all of them just because airplane mode is on.

Change-Id: Iff23d58304be15cfbcd0a69ab1f6f6352132f776
Signed-off-by: rk134 <rahul.kantrapally@gmail.com>